# docs/pro/openapi-v1 2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Game Panel PRO API",
    "version": "2.0.49",
    "description": "Game Panel PRO integration API. Scoped expiring tokens inherit current owner permissions. Native backup creation uses durable admission and never automatically re-dispatches an uncertain request."
  },
  "security": [
    {
      "apiToken": []
    }
  ],
  "paths": {
    "/api/v1/servers": {
      "get": {
        "operationId": "listServers",
        "summary": "List accessible servers (servers.read)",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          },
          {
            "name": "after",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Page ordered by global server ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "nextCursor",
                    "requestId"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Server"
                      }
                    },
                    "nextCursor": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/servers/{id}": {
      "get": {
        "operationId": "getServer",
        "summary": "Read a scoped server (servers.read)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server inventory record",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "requestId"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Server"
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/servers/{id}/resources": {
      "get": {
        "operationId": "getServerResources",
        "summary": "Read the latest resource snapshot (resources.read)",
        "description": "observedAt is the time the runtime assembled the snapshot. Disk measurements are independently cached. Missing or expired samples return resources:null, not zero. Remote runtimes require the updated agent.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          },
          "200": {
            "description": "CPU in vCPU cores, memory/disk in bytes, percentages relative to assigned limits",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "requestId"
                  ],
                  "properties": {
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "observedAt",
                        "resources"
                      ],
                      "properties": {
                        "observedAt": {
                          "type": [
                            "string",
                            "null"
                          ],
                          "format": "date-time"
                        },
                        "resources": {
                          "type": [
                            "object",
                            "null"
                          ],
                          "additionalProperties": false,
                          "properties": {
                            "cpuCores": {
                              "type": [
                                "number",
                                "null"
                              ],
                              "minimum": 0
                            },
                            "cpuLimitCores": {
                              "type": [
                                "number",
                                "null"
                              ],
                              "minimum": 0
                            },
                            "cpuLimitPercent": {
                              "type": [
                                "number",
                                "null"
                              ],
                              "minimum": 0
                            },
                            "memoryBytes": {
                              "type": [
                                "number",
                                "null"
                              ],
                              "minimum": 0
                            },
                            "memoryLimitBytes": {
                              "type": [
                                "number",
                                "null"
                              ],
                              "minimum": 0
                            },
                            "memoryLimitPercent": {
                              "type": [
                                "number",
                                "null"
                              ],
                              "minimum": 0
                            },
                            "diskBytes": {
                              "type": [
                                "number",
                                "null"
                              ],
                              "minimum": 0
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/servers/{id}/backups": {
      "get": {
        "operationId": "listBackups",
        "summary": "List backup metadata (backups.read)",
        "description": "Names use code-point ordering. Scope and current owner permission backups.read are both required. Inventory is bounded to 2000 entries; oversized inventory returns 503.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          },
          {
            "name": "after",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 255
            }
          }
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          },
          "200": {
            "description": "Backup metadata page",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "requestId",
                    "nextCursor"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Backup"
                      }
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "nextCursor": {
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "post": {
        "operationId": "createNativeBackup",
        "summary": "Create a named Native backup (backups.create)",
        "description": "Idempotency keys are scoped to the token and retained across restart. Reuse the same key after a lost response. Different effective name or server with the same key returns 409. An uncertain dispatch is never automatically retried. A 202 is admission, not completion. Read Location with operations.read to check the final result.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{16,128}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Optional label; up to 64 letters, numbers, spaces, dots, hyphens or underscores, beginning with a letter or number; at most 128 UTF-8 bytes. Empty generates a runtime name."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          },
          "202": {
            "description": "Operation admitted or replayed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "requestId"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Operation"
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            },
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                },
                "description": "Operation status endpoint"
              },
              "Idempotency-Replayed": {
                "schema": {
                  "type": "string"
                },
                "description": "true when the same admission was returned"
              }
            }
          },
          "409": {
            "description": "Different payload for the same key, or uncertain dispatch. outcome_uncertain includes data with the operation ID; inspect panel history before using a new key.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error",
                    "requestId"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      }
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Operation"
                    }
                  }
                }
              }
            }
          },
          "413": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/operations/{id}": {
      "get": {
        "operationId": "getOperation",
        "summary": "Read a backup operation (operations.read)",
        "description": "Requires the same owner, current backups.create permission and a token scoped to this server. Runtime replacement returns 409; offline runtime returns 503 instead of an invented result. Internal output, paths and runtime IDs are omitted.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          },
          "200": {
            "description": "Current runtime operation status or explicit uncertain dispatch",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "requestId"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Operation"
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "apiToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "gpp_ token"
      }
    },
    "schemas": {
      "Server": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "number",
          "name",
          "provider",
          "status",
          "observedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "number": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "observedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Backup": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name",
          "kind",
          "sizeBytes",
          "modifiedAt"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": [
              "file",
              "directory"
            ]
          },
          "sizeBytes": {
            "type": [
              "number",
              "null"
            ],
            "minimum": 0
          },
          "modifiedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
      "Operation": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "serverId",
          "status",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "serverId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "accepted",
              "dispatching",
              "uncertain",
              "running",
              "completed",
              "failed",
              "interrupted"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "startedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "completedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      }
    },
    "responses": {
      "Error": {
        "description": "Structured error; requestId also appears in X-Request-ID",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "error",
                "requestId"
              ],
              "properties": {
                "error": {
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "requestId": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        }
      }
    }
  }
}